Add a script to fake-delete a user account on SP
This script "deletes" an account by overwriting their hashed password
value with a string indicating the account was killed by Support.

The R/W password must be entered by the calling user, and is read by
raw_input. In other words it is shown on the screen. Be careful.

A fix for an old TODO in getuser is also provided; let's remove
one further SQL injection vector.

On the plus side, this program does include a molly-guard.

http://www.catb.org/~esr/jargon/html/M/molly-guard.html

AMENDED: fixed a tab.

Change-Id: I3a5890b8d16510f1825541f1e5594b72aee7fbe0