Fix bugs in getHttpConfiguration.
Previously the configuration URL would not be pulled from the default
configuration sources if it was not passed by the caller to
Server.initialize(). This commit fixes that bug.

Also, I was passing the URL to PropertiesConfiguration as a String. Now
I am properly passing the URL as a URL.

initialize() should now appropriately throw if the URL source is somehow
invalid (as per the docs... I have not been able to test this yet).

Change-Id: I477d9a4e0fb51b9f7fe785aa06362d643ce14862