Add UI option to enable/disable sync history.
This adds an option to the preferences panel to enable/disable
sync history UI. Option is added to the client and team server UIs.

This stores the sync history setting in the config db. The
CfgStoragePolicy class is maintained to return a cached setting
to the various places we need the value.

Note that the previous (undocumented) mechanism, a NOHISTORY
file, is no longer supported.

User-visible strings that said "sync history" are updated to
read "Sync History", except where it interfered with
conversational sentences.

AMENDED: removed use of DynamicStringProperty. This cannot be overridden.
AMENDED: gerrit complained about merging. Rebased 2.

Change-Id: I4f68c20caf92e3f20c62e648374e9f7ec5eebc01