Copy logger configuration files into the release.
The logger configuration files must appear in the root of a classpath
element (in this case, in the root of aerofs.jar).

This is an evil hack that does the trick. It is the minimal change
to unbreak this, and will be revisited when Allen touches the build
system (or immediately afterward).

Change-Id: I09976a4a83aff6551cbd4598812d57f9fe9cb43e