Fix seed-file corner case
Reinstalling with a different userid would potentially
break sharing as the SID is derived from the OID of the
folder being shared and that OID would be restored form
the seed file.

Associating seed files to the SID from which they are
generated fixes this problem while preserving the
expected behavior of seed files for external shared
folders.

Change-Id: I1019e1c5f094584789360f58390acd56a4b6427d