AE refactor: step 1
Use a new table to keep track of the set of devices
contributing ticks within each store.

Tradeoff:
 * slight increase in code complexity
 * clearer table roles
 * reduced DB size (small table vs index on large table)
 * faster query
 * in-memory cache

getAllVersionDIDs used to be a bottleneck in building
GetVersReply on stores with large numbers of objects.
The creation of an index specifically for it also caused
a significant DB increase (the version subsystem and its
many tables/indices are the biggest contributors to the
size of the core DB).

Change-Id: I7f48fbb14edc393e99c2efe185ddcc1c0aa6e236