Fix handling of properties files for staging
The needs for the client and servlets differ with respect to the location
of properties files. In both cases, the artifacts need to wind up in
the classpath.

For staging clients using exploded approot, the classpath destination
and root destination differ (approot/bin and approot).

For staging servlets, the root and classpath destinations are the same.

This checkin updates the copy_mode_and_product_resources mechanism
to support both use cases.

Change-Id: Ifb1595ab4cb8cee27b266db400c93a1be31770a5