Suppress stack trace when writing a defect fails
The stack trace in SV's log is not interesting when a defect cannot
be written to the disk (for example when the disk is full). The
log is clear enough, we don't need the invocation context.

This should reduce the overhead associated with the /data partition
being full on this server.

Change-Id: Iaf644cb81a754fd3dfc686e2e40b347c62478aee