Add ejabberd tasks to reset metadata.
Ejabberd is stupid and locks to a specific hostname when it first runs.
Since it runs during the bake process, it will not run again unless you
blow away the metadata. Add a bootstrap task for this, and rename the
old ejabberd task.

Small fix to the pull installers script as well, but more fixes will be
needed here later.

Change-Id: I0c0c3b3f837df1f6d367c6c106ba75994aaf4b99