Fix many bugs in private deployment Linux install
This was *clearly* never tested before now.  Issues fixed:

  - Missing whitespace breaks [ invocation (no ] found)
  - Lack of quoting breaks config handling
  - Silent introduction of dependency on wget (not included by default
    in many distros!)
  - Fails to actually produce correct download URL

Some of these might have been hard to catch if you don't have a config
server up and running to test against, but the bash syntax error says
that you never so much as tested that branch of the code.

I Am Not Impressed.

Change-Id: I325f8ae29d7b69e3e3c478d8094db35e0221d143