Fix some syncdet script issues.
Noticeably, this changes the command-line parsing of systest-setup for team server
tests. The --include-ts options are gone, replaced with
    --ts {S3, LINKED, LOCAL}

which enables team server for one actor, with the given storage type.

This will require a matching CI change to match. This fixes the issue that Team Server
tests could not run on the isolated LAN actors.

Switched the convert-to-lowercase to use a more idiomatic tr convention.
I have no idea how this works on the CI clusters, unquoted square brackets as
used in the tr invocations are Just Plain Broken. See tr docs if you don't believe me.

This was blocking syncdet on my local vagrant environment.

If OSX used bash-4.* then we could just use ${VAR,,} as a lower-case
parameter expansion... Oh well.

AMENDED: fixed idiomatic python

Change-Id: I9ecccff3f7c182285fcb3d7e518c23763662a80f