Defect: Gather disk usage for physical roots
Until now we only checked disk usage and fs type for the rtroot
which was fine because most users put the AeroFS folder on the
same partition and on Unix we have the output of df to help us.

However on the Windows we need to manually gather disk usage
and fs type for all physical roots to help diagnose disk full
problems.

Change-Id: Idd71b4e93918eb30f4ddf47a58d698bb169e984a