Disable Metrics reporting
This simply disables the yammer Metrics reporter, since metrics currently
has no topic owner. This implementation is generating traffic that we
don't review.

Metrics will be re-added when someone picks this up as an area of
interest.

AMENDED: Removed rocklog.Metrics and rocklog.RockLogMessage (the
        abstract superclass).

        com.aerofs.lib.rocklog.Defect now stands alone, for better or worse.

AMENDED: short-circuited the metrics/ path on the rocklog python server.
        We can remove this later, but it seems unfortunate to return errors
        for clients that have not updated to avoid sending metrics.

Change-Id: Ida99498f84e69511a8c4356b115c0ee8f9041859