Remove "cannot invite self" error checking.
Yuri pointed out he cannot add himself to a Beth's shared folder even though he
is an admin. He got a "cannot invite self" error message. This is a relic of the
past - from back in the days when the only person who could invite new users to
a share was the creator of the share. Now, if you're an admin, you can invite
people to a share as long as they're not already a member (or at least that's
how the system should behave). The "is this sharer allowed to edit the ACL" and
the "is the sharee already in the share" checks already exist, so my job was
easy - just remove all "cannot invite self" checks, and Yuri's problem will be
solved.

N.B. by admin in this context I mean team admin not folder admin.

Change-Id: I69afb7a8afd7d2452b24b1850410f2ab0e413753