Kill and start processes in syncdet more reliably.
Using pid files in cygwin/windows is flaky at best, since we often get
access to the cygwin pid, but can only kill the process with the winpid.

Processes are now searched for by name using tasklist or ps, depending on
the system. The output is grep'd for the process name to find the pid, and
the process is killed with os.kill(pid, SIGKILL) or taskill /F /PID winpid.

In addition, processes sometimes take a little time to die. There are cases
where both the cli and the daemon will be killed, but the cli will have time
to restart the daemon, so the daemon remains in the end. Killing a process now
blocks until the process no longer shows up in the task list.

Finally, starting the daemon now blocks until it has a heartbeat, so that syncdet
tests can't make rpc calls while the daemon is still indexing (which is bad!).

Change-Id: Ifa9b9b8df4c43c4a2ed4dbe691285fcd2df7fec4