Bound retries on UPDATE_IN_PROGRESS
When a peer is uploading a file it will check for local
modifications after each chunk is uploaded and abort the
transfer if the file changes.

In such cases we usually want to retry the download fairly
quickly to pick up the new changes.

However it is possible that some file end up in a state
of permanent changes (typically log files...) and some
seemingly innocuous files have been observed to cause
a similar pattern of transfer abortions for a reason
that remains to be clarified.

If a peer starts downloading 5 such unstable files, a
nosync would result as the downloads would fail and be
retried every 3s

To prevent such a livelock, the number of immediate
retries is now bounded. Most updates will therefore
result in a swift and successful retry and the rare
problematic files will no longer inhibit forward
progress of the collector.

Change-Id: Ifa68737bc3c612a49de2a7c2fc6516ac06c05a12