DefaultDaemonMonitor: Fix race condition in daemon startup code
After starting daemon, cli tries to connect to the ritual notification server.
Previously, if this failed, cli would assume that the daemon was dead and start
a new daemon (which loops, spawning ALL the daemons). Now, if connection to rns
fails, check for a heartbeat to make sure the daemon is in fact gone. If the
heartbeat returns ExIndexing, then the daemon is alive and has likely not yet
started the rns, so pause and retry.

AMENDED: jP: Just an alternate implementation that doesn't worry
about indexing/updating states and treats connect() as a timeout-able
encapsulated action.

AMENDED: jP2: Ignore those build files...and that Math.fail().

Change-Id: I2a36a9a78a41e2484226384e0c8e2862de48d998