Fix OOM in DPUT
Slightly more aggressive cleanup of immigrant ticks
to avoid O(ticks) memory usage which has been observed
to cause OOMs in the wild.

Also add safeguards to avoid removing local ticks
(which would not be restored through AntiEntropy
hence leading to a possible no-sync if expelling
and readmitting a shared folder before the dput
was run).

Although re-running the updated DPUT would not be
necessary, local prod mixup recently introduced fresh
ghost KMLs in the "Air Computing Team" folder so given
that recent network changes already bumped the core
magic it was decided that bumping it again was acceptable.

In the longer term, it might be necessary to devise
a way to incrementally cleanup old KMLs without
requiring a magic bump as there are still a couple
of valid scenarios that can generate such ghosts.

Change-Id: I864986d3b7ab7a90fad11366c2d7693d4c48c406