Don't let pip upgrade everything every time.
This adds an explicit requirements.txt with current versions for
artifacts we care about on webadmin. Previously we used --upgrade
which upgraded everything to the latest on PyPI.

This broke the world when setup-tools-0.7 became the new hotness;
it does not play nice with something in our world.

Tested by blowing away the virtualenv on ci.webadmin and running
the included postinst script by hand; it appears to work.

AMENDED: changed versions to match production deploy

Change-Id: I6dbff8038f8c95c6c3511b33ec7c6fae26d0275a