Fix uwsgi puppet manifest
UWSGI has a broken status command - it requires a conf file argument.
Puppet invokes "status" with no args, uwsgi prints an error but puppet
doesn't realize it is an error, and it never starts/restarts uwsgi.

This fixes the uwsgi requirement in webadmin module.

For some reason this started breaking CI.

Change-Id: I4fe058382c70f53eb4dd7e25de2429355e167c41