Catch a possible zephyr NIO error and terminate
We already check for failures from nio accept(). However, rarely
NIO may pass us a socket that is not actually connected. This
results in NPE when we try to print the remote socket address.

Our handling of unexpected exceptions in accept is weird - result is
we stop the channel but don't drop out of the dispatcher loop.
So, Zephyr winds up in a zombie-ish state.

This change explicitly checks the connected-ness of new sockets,
and treats that situation as if accept() had thrown IOE.

(To simulate this NPE behavior, try adding a call to
sc.socket().close() after accept() and before getRemoteSocketAddress()
is called...)

Change-Id: Ibab69cf39e527c8ae42968f72ffe7f90a9051007