{Windows,OSX}Notifier: remove deadlock.
We used to use the JNotify-provided JNotifyWin32Adapter, but since we
need to be able to queue core events without holding a JNotify lock,
we need to roll our own.

Fortunately, we can still rely on the Windows recursive watching to do
the right thing, so this code can be substantially simpler than the
LinuxNotifier, which inspired the deadlock-avoiding strategy taken here.

On the OSX side, we again need to ensure that enqueuing events into the
core does not require holding a Notifier lock.

Change-Id: Ie7432d05c883bc48b240496c87082cb8f0d5da10