Fix zero-byte startup.log droppings
AMENDED: removed an unused String from Main.

Ok, so logback configuration is pretty annoying. The XML was starting
to get complex and the eager-loading appender strategy means we were
leaving zero-byte files called startup.log. This in turn broke
code signing on OSX, kinda.

This checkin breaks the client logback.xml down to a NOP appender
in the xml file. On client application startup, we create a programmatic
configuration context and appender (incorporating the program name
and the RTRoot, which is why we can't do this in xml).

Result is that users should see the same log behavior as always, but
without the startup.log file left behind.

Change-Id: Ia05d4332e52d066755784911579c8f522d98e025