Move growlbindings jar to all OSes
The way the source is written and the way Java works, the growlbindings
jar is required at build-time for all platforms.  I'd rather ship a 9Kb
jar on all platforms than deal with the madness of conditionally
including a jar needed at build time and specified in a manifest.

This is another step toward Linux build agents.

Change-Id: I6adf22f672734b05b6747ee175d031adbcc4f26b