Refactor all setup logic for single and multiuser
No user-visible changes, but a big refactoring of the classes
involved in setting up a client. This is foundational for
adding OpenId support.

 - No more use of Controller for setup/signin/install

 - Removed unused classes

 - Reworked SetupModel to hold all context for the setup. It is
 configured with actors for single or multiuser install, and can
 be updated to support different signin flows.

 - SetupModel is now the public face of the Setup class,
 which continues to be package-private.

 - Setup only signs in to SP once (was twice)

 - Merged out an abstract class with one child

 - fixed a build.xml bug that broke local-variable-debugging

Tested Team Server and singleuser client by hand in CLI and GUI,
and it appears to work...

AMENDED: squashed one CLI bug that snuck through.
AMENDED: resolved review comments

Change-Id: I00d23d04cd134534f8fab261f223d9819910a94e