Initial implementation of OpenId authorization.
This adds a new servlet to SP and all the GUI code that goes along
with it. The identity servlet forwards auth requests and response to
an OpenId Provider to authenticate the end-user.

AMENDED: updated with review comments from Alex & Allen, and rebased.

The goal is to be flexible enough to handle Google, MyOpenId, and
hopefully even Bloomberg's OpenId implementation, which all
deal with the spec in wildly varying ways.

The servlet and GUI code is only enabled when local-production mode
when openid.service.enabled is set true.

This has been tested with client, team server, and CLI.

Notes to reviewers:
    The GUI code and server code are more or less disjoint; the boundary
    in sp.proto is the only interaction between these components.

    The entire login sequence is somewhat tricky; see
        docs/design/openid_auth_sequence.png

TODO:
    configuration options are documented in LibParam; they need
    to be documented elsewhere (for a sysadmin)

    GUIs are functional but sucky.

Change-Id: I63311ed632cf87befb90583dc8d392e6bf538dfb