Fixed a regression.
Symptoms: File conflicts are not shown in the tray icon and menu
when the client restarts even if there are conflicts.

Root cause: The intended design is that RitualNotifier notifies
all listeners when a RitualNotificationClient connects to a
RitualNotificationServer. And NotificationService should register
as a listener so it can send snapshots to RNCs when they connect.
The NotificationService currently fails to register as a listener.

Change-Id: Ie8e64ea6c77b62d734b8135eddcbf9bad01fca3a