DynamicConfig Removal: Phase I. Read configuration into Properties instead.
The first phase of removing DynamicConfig. During intitialization we replace
DynamicConfig with Java Properties. We then use the Properties to intialize
DynamicConfig so that we can refactor out actual usages of DynamicProperties
in a future review. Some important changes are:

* BaseParam now uses SimplePropertySource in all cases (Android, Desktop
client and Servers). In a future review BaseParam won't use a PropertySoruce
at all.
* When defining configuration properties we now have the policy that last
setter wins.
* When a property is multiply defined we log a warning and proceed with the
last setter winning.

Closes #126