resources: pre-sign Finder extension
To detect if we need to reload the Finder extension, we compare the hash
of the one on disk with that of the previous release.  If they differ,
we kill Finder.

Codesigning modifies the actual executable that you are signing.  As
a result, resigning the AeroFSFinderExtension .osax changes the binary,
which prompts AeroFS to kill Finder on the update, even if there are no
functional changes.

This commit signs AeroFSFinderExtension.osax, checks in the signed
artifact (along with its matching CodeResources) and disables
automatically signing the finder extension on future deploys.

Closes #41