Added a feature to show server status on tray icon.
Server status is defined as whether the client is
currently connected to Verkehr.

When ritual notification client connects to ritual
notification server, the current server status is
sent to the client as a part of the snapshot.
Whenever server status is changed, the change is
propagated to the client via ritual notification.