Add a DumbAss option to OpenId server config
So, there exist such weird servers in the world that they don't support
Diffie-Helman key exchange. This is forbidden by the spec but here I am
trying to talk to just such a server.

This checkin adds a LibParam option for a Dumb Association mode. (This is
the nomenclature from the standard, thank-you-very-much).

Hence the abbreviation.

The server config option is
    openid.idp.endpoint.stateful
default is true, like a gentleman.

Change-Id: If051638d6dd9957fb57b581dc6bd09a5102a73ac