Android: Fix deployment problem where it would fail if map file already existed.
When saving proguard maps to SV and Rocklog, scp would fail if another user
had already saved the maps (ie if anyone had deployed this version of Android
before you, your deployment would fail when uploading Proguard maps). To fix
this we allow SCP to fail and not stop the deployment process.