LibParam: Make EnterpriseConfig.IS_ENTERPRISE_DEPLOYMENT non final.
It's desireable to change the value of IS_ENTERPRISE_DEPLOYMENT in tests. We
currently achieve this by leveraging DynamicConfig. This review allows the
IS_ENTERPRISE_DEPLOYMENT property to be changed without reaching through
DynamicConfig.

Some justification for this: Currently the use of final doesn't keep properties
from changing, it just forces you to reach through DynamicConfig to change
them. Removing this reliance on DynamicConfig to bypass the unchangeability of
a final field will allow for simpler removal of DynamicConfig.