Fixes bug: After updating OSX users see "AeroFS is still running" on log in/out.
After updating on OSX, we restart aerofs with 'nohup', which prevents OSX from
stopping AeroFS when the user logs out. This means that on the next login (if
there isn't a shutdown) the user sees the "AeroFS is still running" dialog,
and the new instance of AeroFS fails to start. This is particularly annoying
because the tray icon doesn't show up, but AeroFS is still running and syncs
files.

'open -a' starts AeroFS while allowing OSX to stop AeroFS on logout.

Closes #168.