OSX codesigning: ignore site-config.properties
In general, adding unexpected files to a signed package which includes
Resources/* in the resource rules breaks the codesigning seal.  Then,
when we add site-config.properties, we break the signature.  Sad day.

To get around this, we can specify that the site-config.properties file
should not be considered part of the seal.  Then, the application bundle
will verify successfully, regardless of the file's presence or absence.