packaging: add a tool for caching pip packages
Right now, we download Python packages from PyPI at install time, which
is problematic because:

1) It's slow
2) It's not repeatable
3) PyPI can have downtime, which wrecks our flow.

pip-prefetch.sh is a shell script that will fetch all the source
packages given in a requirements file and store them in a folder.
Additionally, if the requirements file hasn't changed since the last
time that cache was fetched, it'll just use that one.

Combined with some modifications to the builder scripts, we can make it
easy for our debs to provide their own dependencies and install
repeatably, even offline.