Initial implementation of bifrost
Bifrost is an OAuth frontend that delegates user authentication (username/password)
to an SP call. In other words, we generate an OAuth token if the user proves they
can log in to SP.

Much work still to be done. This is a first milestone, which is that the service
_can_ run and talk to a MySQL database.

AMENDED: added schema sql
AMENDED: refactored to use restless instead of dropwizard
AMENDED: integration with netty/restless; added initial unit test

Change-Id: I2042dad4d424371496958f953d769e2d800e7ae9