pip-prefetch.sh: do stricter check of cache
Something happened on Yuri's laptop that resulted in him building a
package with an incomplete sdist.  pip unhelpfully returned 0, so
packaging proceeded.

This change adds some sanity checking to the pip-prefetch script to
verify that some file with a probably-matching name exists for each of
the packages specified in the given requirements.txt to try to tease out
this sort of failure even earlier.

Change-Id: I6dcf651d985aef7c7fda68ac9b55285e13b26aa0