private-deployment: Remove unneeded mysql users
Puppet's mysql module has a class for removing some of the
default-created users that aren't likely to be needed in practice.
Among others, this removes the ''@'localhost' user, which previously
caused tomcat to be unable to login, since 'aerofs_sp'@'localhost'
matched the ''@'localhost', but did not have a grant for the aerofs_sp
table.

Using this class allows us to drop the bootstrap task that was added to
work around this issue, so away it goes.

Change-Id: I1838a8736a305810c22bbfdfaf475bbb8e810ef4