syncdet_test: use newci-style enterprise installation and config
This commit rewrites the way our syncdet_test code handles installation
and process management to use private cloud installers. It also removes
code that is obsolete in the new CI environment.

N.B. "Old" CI will not be able to function with this commit in master,
so this commit points verify_ci_status at newci

Change-Id: I6ea926ed43c811fd9fa9a450ebd938650ce9c398