Web login supports openid/credential in one screen
This is ongoing development for the enterprise deployment using OpenID;
the external users will be offered the chanceto log in with an existing
SP credential.

The refactoring to get here resulted in deleting a bunch of near-duplicate
code; the entire openid renderer stuff has been fixed up, and the
login_credential view is gone as well.

NOTE the web layout is ugly! Hoping someone will help me understand
what all the div classes/styles mean and lay out the buttons and hints
in a way that human beings will respect.

AMENDED: L.product() in place of constant string; use is_enterprise_deployment
Change-Id: I56643bd829feb7d6831bfd860f632824f79877e8