Do not allow pw reset for externally-managed users
This disallows the password-reset flow for users that are in enterprise mode,
and have an LDAP or OpenId authenticator, and who are internal users. Wow, right?

A new exception type, ExCannotResetPassword, is added and reported via the web UI.

The PasswordManagement class now takes a UserFilter to enable the above.

Change-Id: I2826bef1b7a68d280c968dbe545505d2ca79ec35