Fix a regression in OpenId error handling
When I merged the signin dialogs, a bug was introduced that OpenID
timeout was reported to the user as "bad email address or password".

This adds a new wirable exception type, ExExternalAuthFailure to indicate
a non-specific failure from some external authenticator. This is used today
to indicate an attempt to use a bogus, already-validated, or timed-out nonce.

At the same time this fixes a one-line bug in the OpenID signin web UI,
which is to use the openid.service.identifier in the signin button.

AMENDED: Updated comments re: thrown exception
Change-Id: I8790e7df3620d95b97fa46e1c9d3e5ad6c4ee58c