SPService: reduce info-level log lines to trace/debug
On every getACL call, if an ACL had been updated, SP logs all the members of all
the folders of which the caller was a member. So, if a share had N users, each
would make a call triggering N lines at info-level. These were changed to be at
trace-level. One log line was added at info-level for each getACL call.

Change-Id: I48aa2a139ed44a0a91ed5f726e12721d01d09a15