Fixed an issue with environment variable substitution.
- Also removed unused imports.
- Dropping support for "~" on Windows.

The problem we ran into at Bloomberg was as follows: the environment
variable $HOME was not defined on a Windows install, so when the code path
to substitute environment variable on Windows is executed, we end up with
a NullPointerException.

There are couple problems with the technical implementation:
- The previous impl. assumed that $HOME will always be defined, it is not
the case on actual Windows install.
- The previous impl. is indiscriminately replacing ~ regardless of whether
it is necessary.
- The previous impl. did not guard against $HOME being undefined.

The reason why we have not run into this in testing was that when I do
manual testing, I launched the app via the shell on a Windows VM. The shell
supplies ${HOME}, which I mistakenly thought was supplied by the JVM
instead, and everything worked.

Catching errors like this requires doing a local package, repackage, and
install via an installer.

Change-Id: I61be716c34d3642b998b643a9e926a4b4eb20cc5