Fix web setup login security issue.
Problem is as follows: it is possible for an attacker to forge a cookie
such that they can access admin-only pages without properly signing in.

Tech details: we are storing the auth group in the cookie. We did this
to avoid round trips to SP. This worked previously because all sensitive
data had to be obtained from SP using a tomcat token anyway. On logout the
token is revoked so attackers cannot forge a garbage cookie to steal our
data - all was well. However, with the new pages that do not interact
with SP, it is possible to forge a malicious cookie, since auth is not
checked with the SP in this case.

One possible solution: always load auth from SP before rendering. The
result is some unnecessary round trips to SP when we're already querying
SP anyway. However, this is the safest solution. Requiring web devs to
think about reloading auth is a bit scary, but if we're really opposed
to the extra round trips, that is a viable solution as well.

Change-Id: Ibcd94bed6ea7c80b097c746dbfe7ef5dfbca42f4