Reverted automatic line-wrapping in CLI.
The root cause of a number of text-wrapping defects is an earlier
decision to make the CLI automatically wrap #allthethings.

The implementation uses org.apache.common.lang.WordUtil.wrap(), which
doesn't take "\n" in the string into account when it wraps lines.

In retrospect, the decision to automatically wrap #allthethings is a
mistake. This commit is to revert that decision and undo all line wrapping
in the CLI.

Change-Id: I89ee83284315d9b0f4ae11fd647b0e3c76f879c7