Modular nginx configs
The purpose of this change is to simplify configuration management and
the general management of a growing set of services and paths.

With this change, deb packages providing backends for which nginx is
expected to reverse-proxy also ship a configuration file with a location
stanza in /etc/nginx/backends-available/.

This allows for separation of frontend and backend traffic, which allows
us to let only the frontend be concerned with which SSL cert is being
used, and lets the backend be concerned only with the service that it
provides at the appropriate path in the REST hierarchy.

This also, in due time, will allow for making our production deployment
to more closely mimic our private deployment.  By using that same
internal ports for the same purposes everywhere, we can have more
flexibility and less complexity in deployments.  For now, though, we
only use these new configurations in our private deployments (though
they will exist on our public deployment servers).

This change also makes it easy to move the config service (or at least
the public-facing third of it) to simply reside at a particular path in
the unified REST hierarchy.  This allows us to remove the need for port
5435 to be exposed to the world.

Change-Id: I106b6cf74cb336a90752e3e43afb06a7e2970fc5