Fix "verkehr errors" in local prod.
The sp servlet was not starting up properly because static
initialization of ldap parameters was failing. This was happening due to
the templates that weihan recently added. Our ldap config class
unfortunately does not handle itself gracefully when given a bunch of
empty strings.

TODO: refactor ldap so that we don't need sane defaults in these
specific locations. For now, just add the defaults we need to survive.

Change-Id: Id0f0c42ad847ef377ed6b8d0b48f457b6f40f43f