Use distutils for python-lib, make sdist
It'd be cleaner if aerofs-py-lib installed the same way as the rest of
our packages.  This commit makes "make web" produce an sdist package for
python-lib which gets installed with the rest of the packages in the
postinst script, rather than copying in the whole source tree for
python-lib.

Independently-distributable packages are good for the soul.

Change-Id: Ie502b73e0db692204a53019dfe54762071351f97