Package web as "aerofs-web"; use sdist for install
This makes our web module get packaged and installed through the same
flow that we use for all python packages.  aerofs-web is now just
another python module that we drop in the sdist folder and install as we
would any other.

This commit changes a file provided by puppet (production.ini); as a
result, this commit breaks in-place upgradability.

Change-Id: Ibc38e1f9d4bedf2e9b94614ce8ba297fcff5ab5d