Detect core db tampering
Restoring a backup or using a migration tool to copy
user data from a device to another has the potential
to wreak the entire distributed system by corrupting
the tick space.

To prevent that we attempt to detect such occurences
by relying on the fact that the FID of the core DB is
expected to remain constant during normal use but will
change if migration/restore happens.

IF such an occurence is detected the user is presented
with a dialog offering to reinstall (default) or force
launch (unsafe).

Change-Id: I9dae1831991ee0d783852f837c4b7f6edd3ea2a7