SP support for device-authorization nonces
This provides the necessary hook for the web to generate and return
a unique nonce that can be used to authorize one device. The request
for a device authorization nonce requires User-level authentication.

Nonces currently last for 3 minutes; TODO: make this configurable?

AMENDED: updated based on some review feedback
AMENDED: Better naming embiggens the smallest concept

Change-Id: I41b3d58d71c69028a5807fe9749203f6b78f5796