Add device-authorization-nonce token exchange
This checkin allows a new OAuth token request semantic.

Users request a token using the "authorization_code" grant type
(which is standard). They submit an SP-generated device authorization nonce
as the access code, and indicate that we should interpret the request this way
by attaching a parameter "code_type="device_authorization".

The effect is that we will present the device-authorization nonce to SP,
and save the returned UserID values in the access token record.

TODO: commit the unit tests
Change-Id: If7382bd70dbf1aa6e9c78dcd0bf2df0a082e6b21