licensing unit tests
A significant refactoring was required to allow for injection of a
separate gpg homedir.  Note that the keys under test_keys will not be
trusted by prod, and that the production ones will not be accepted for
unit tests.

This allows us to check in some test license files that will be valid
for a very long time, for instance.

Change-Id: Ie78ed523b38d10e4defacf6bc789600ef1af28f3