Fix a pile of unit tests I broke
It turns out User.save() gets used by a bunch of unit tests!  And they
expect to succeed without license things getting in the way!  How
surprising!

We now inject a License into User.Factory, which we can mock in tests.
The mocked License is valid, and allows an excessive number of users.

Change-Id: If7f31ff44ff52ec6ddb14c2d59024b17cbc76600