web: don't import *
Python will hoist *all* dependencies of a package when you import *,
including everything that that package imported.  This can result in
code that happens to work but does not declare its dependencies very
clearly, which makes changes fragile.

It's better to be explicit about your imports to avoid accidentally
winding up with more than you expected in your namespace.

Change-Id: I68c6e60dc668689b3ca50b42cc98685edae3770f