Remove SCrypt from LDAP and web paths
SCrypt'ing the user credential will now be done server-side. This checkin
includes the server implementation as well as converting two cases to
use the new way:
 - clients using "external_credential" authentication, i.e. LDAP
 - web signin
 - password reset

All python use of SCrypt is now gone.

The hybrid-cloud and local-credential clients have not been updated, though
this work can occur as soon as we have a free minute, and will be a net
reduction in code.

This work is in service of simplification as well as paving the way for
better credential-salting in the future.

New unit tests have been added for the credential paths. Tests for legacy
signin methods have also been added (and maintained where they existed).

AMENDED: updated encoding issues per Drew's review
AMENDED: plus jgray
AMENDED: and your little dog too

Change-Id: I0c9e2f9528bc657470362843ea467ee0d43d8ecc