Fix corner cases in aliasing and BlockStorage
BlockStorage should succeed in deleting missing files.

Aliasing could in some cases contribute to the appearance of
such missing files because in the fast path it omitted
physical moves which was mostly fine, except for cases where
the physical objects where addressed by SOID, in which case
shit just broke...

This issue was detectedi in an audit of our TeamServer logs
yesterday. Unfortunately there is no easy way to detect and
automatically recover from this issue. On the brighter side
the likelihood of actual users hitting this issue is fairly
small.

Change-Id: I4e1c68f522061f2b63ef28386e49f18e10c6a787