Rework Authenticator paths for consistency
This makes a few changes to the IAuthenticator type.

First, the interface is renamed to IAuthority - LDAPAuthority, LocalAuthority,
and (new) OpenIdAuthority each model an identity authority. Someone who can
tell you if a user is real, and what kind of account they have.

The class that exposes this to other functions is called Authenticator,
and it is a concrete class that delegates behavior to one or more Authority
instances.

The Authenticator interface is fixed up to allow shared-folder rules, authorization,
password change, and other paths to ask simple questions:

 - does this user have a locally-managed password?

 - is this user internal or external to the organization?

 - does this credential let this user authenticate?

A bunch of new tests are added for identity management, and existing tests are fixed up.

AMENDED: naming and comment changes from @WW

Change-Id: If8f8e923ecfbc22e65150d3d5415ba641a71997e