Delay tag file re-creation
CI highlighted the dangerous consequences of eager re-creation,
which, on Windows could prevent a deletion from completing.

Instead of fixing the tag file as soon as we detect it to be missing
or corrupted we now schedule a core event to do that after a safe
cooldown period (the 6s timneout used in TDB is reused here).

Change-Id: Iee8f91aaabb0a6e8dfa5b5414d6bf71ce4e491b4