Fix template-parsing bug, add OpenId params
If the template specified an integral-type property that
was not included in external.properties, the symptom is
that the python config server would not launch and the whole
world dies.

This fixes that by considering an empty string equivalent
to a null string for the purposes of integral types.

Also, this adds template parameters for:

 - OpenID
 - readonly shared folder rules

This is tested by hand right now. :(

AMENDED: use guava#isNullOrEmpty()

Change-Id: Ie24e0c724d5729ce526f260d03a2bb9600c4cf04