OSX: more NFC/NFD annoyance
We used to convert everything from the OSX Notifier into NFC, since we
expected that encoding everywhere in the system.

After the addition of the Non-Representable-Objects, we expect
everything we deal with to either 1) be representable on the native
filesystem or 2) be present in the NRO folder.  If we received a
notification about a encoding-sensitive file or folder, the OSXNotifier
would detect it convert it to NFC, then pass that to the core, which
would convert the file to an NRO (as NFC is non-representable on OSX),
scan for the NRO, and find it absent, causing the scan session to abort.

Instead, we can make InjectableFile.list() give an platform-accurate
return value, and expect NFD everywhere we use physical paths on OSX.
(If a path has arrived from the Notifier, it's clearly representable.)

This refactor cleans up some OSUtil -> InjectableFile circular
dependencies.  OSUtil can always be safely instantiated.

Change-Id: I91e3dca4a6ac746f3b0374b9c6b641f7eca6c468